[#177543882] add citizen v2
